+++ .gitignore
@@ -5,6 +5,6 @@ coverage
 
 # legacy exports
 _unstable_apply.js
-_unstable_store.js
+_unstable_local.js
 path.js
 tsconfig.tsbuildinfo
