+++ examples/web/App.tsx
@@ -15,12 +15,12 @@ import {
   SanityEncoder,
 } from '@sanity/mutate'
 import {
-  createContentLakeStore,
+  createLocalDataset,
   type ListenerSyncEvent,
   type MutationGroup,
   type RemoteDocumentEvent,
   type SanityMutation,
-} from '@sanity/mutate/_unstable_store'
+} from '@sanity/mutate/_unstable_local'
 import {
   draft,
   type Infer,
@@ -275,7 +275,7 @@ function observe(documentId: string) {
   )
 }
 
-const datastore = createContentLakeStore({
+const datastore = createLocalDataset({
   observe,
   submit: transactions => {
     return from(transactions).pipe(
