+++ src/local/__test__/squashMutations.test.ts
