+++ src/local/__test__/squashNodePatches.test.ts
