+++ src/local/documentMap/applyDocumentMutation.ts
