+++ src/local/documentMap/commit.ts
@@ -1,17 +1,17 @@
 import {type SanityDocumentBase} from '../../mutations/types'
-import {type Dataset} from '../types'
+import {type DocumentMap} from '../types'
 import {type UpdateResult} from './applyMutations'
 
 export function commit<Doc extends SanityDocumentBase>(
   results: UpdateResult<Doc>[],
-  dataset: Dataset<Doc>,
+  documentMap: DocumentMap<Doc>,
 ) {
   results.forEach(result => {
     if (result.status === 'created' || result.status === 'updated') {
-      dataset.set(result.id, result.after)
+      documentMap.set(result.id, result.after)
     }
     if (result.status === 'deleted') {
-      dataset.delete(result.id)
+      documentMap.delete(result.id)
     }
   })
 }
