+++ src/local/optimizations/squashDMPStrings.ts
