+++ src/local/optimizations/squashMutations.ts
