+++ src/local/optimizations/squashNodePatches.ts
